requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.0.3
d==0.2.2
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-computed-property==0.3.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-sslserver==0.20
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.0.0
dnspython==1.16.0
drf-rw-serializers==0.1.1
eventlet==0.25.1
greenlet==0.4.15
idna==2.7
importlib-metadata==0.20
itypes==1.1.0
Jinja2==2.10
kombu==4.6.4
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.0
monotonic==1.5
more-itertools==7.2.0
mysqlclient==1.4.2
oauthlib==3.0.0
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
Pygments==2.3.1
PyJWT==1.6.4
pyquery==1.4.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.2.0
selenium==3.141.0
serpy==0.3.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24
vine==1.3.0
virtualenv==16.0.0
zipp==0.6.0
